+++ src/annotation/src/AnnotationException.php
@@ -1,16 +1,10 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: inhere
- * Date: 2019-02-02
- * Time: 15:00
- */
 
 namespace Swoft\Annotation;
 
 /**
  * Class AnnotationException
- * @package Swoft\Annotation
+ * @since 2.0
  */
 class AnnotationException extends \RuntimeException
 {