+++ src/db/src/Eloquent/Collection.php
@@ -70,7 +70,7 @@ class Collection extends BaseCollection
      *
      * @return bool
      */
-    public function contains($key, $operator = null, $value = null)
+    public function contains($key, $operator = null, $value = null): bool
     {
         if (func_num_args() > 1 || $this->useAsCallable($key)) {
             return parent::contains(...func_get_args());