+++ src/framework/src/Processor/ConsoleProcessor.php
@@ -2,34 +2,44 @@
 
 namespace Swoft\Processor;
 
-use App\Aspect\TestLog;
+use Swoft\Console\Application;
+use Swoft\Console\Bean\Parser\CommandParser;
+use Swoft\Console\Router\Router;
 use Swoft\Http\Server\HttpServer;
 
 /**
  * Console processor
+ * @since 2.0
  */
 class ConsoleProcessor extends Processor
 {
     /**
      * Handle console
+     * @return bool
+     * @throws \ReflectionException
+     * @throws \Swoft\Bean\Exception\ContainerException
      */
     public function handle(): bool
     {
-        if (!$this->application->beforeConfig()) {
+        if (!$this->application->beforeConsole()) {
             return false;
         }
 
-        echo 'console' . PHP_EOL;
-        echo '---------------------' . PHP_EOL;
+        /** @var Router $router */
+        $router = \bean('cliRouter');
 
-        /** @var TestLog $testLog */
-        $testLog = bean('testLog');
-        echo $testLog->log() . PHP_EOL;
+        // Register console routes
+        CommandParser::registerTo($router);
+
+        // Run console application
+        /** @var Application $cliApp */
+        $cliApp = \bean('cliApp');
+        $cliApp->run();
 
         /* @var HttpServer $httpServer */
-        $httpServer = bean('httpServer');
-        $httpServer->start();
+        // $httpServer = bean('httpServer');
+        // $httpServer->start();
 
-        return $this->application->afterConfig();
+        return $this->application->afterConsole();
     }
 }