+++ test/bootstrap.php
@@ -1,5 +1,6 @@
 <?php
 
+// vendor at component dir
 if (file_exists(dirname(__DIR__) . '/vendor/autoload.php')) {
     require dirname(__DIR__) . '/vendor/autoload.php';
     // application's vendor
@@ -7,9 +8,10 @@ if (file_exists(dirname(__DIR__) . '/vendor/autoload.php')) {
     /** @var \Composer\Autoload\ClassLoader $loader */
     $loader = require dirname(__DIR__, 3) . '/autoload.php';
 
-    // load test psr4 config map
-    $componentDir = dirname(__DIR__);
-    $composerData = json_decode(file_get_contents($componentDir . '/composer.json'), true);
+    // need load test psr4 config map
+    $componentDir  = dirname(__DIR__);
+    $componentJson = $componentDir . '/composer.json';
+    $composerData  = json_decode(file_get_contents($componentJson), true);
 
     foreach ($composerData['autoload-dev']['psr-4'] as $prefix => $dir) {
         $loader->addPsr4($prefix, $componentDir . '/' . $dir);